Updated on Review
- Could have used a maven/gradle for dependency management. (Found me
manually downloading the hamcrest-core jar for running the test cases,
since it wasn't checked in part of the repo)
- Used try-catch with some empty 'catch' blocks. And in most of the
other 'catch' blocks, prints error to the console. This being a CLI app,
this doesn't seem big, but others (like accessing via REST API) would
need a different way of error handling.
- Hasn't modeled the Devices (Computer and repeater) and treated them as
strings. Has a 'Node' Object as a POJO.
- Namings could be better. Ex. CommNetwork, NetworkCmdExecutor